// src/enums/CurrencyValueType.ts
export enum CurrencyValueType {
	EXCHANGE_RATE_NUMBER = "Broj tečajnice",
	DATE_OF_APPLICATION = "Datum primjene",
	COUNTRY = "Država",
	UNIT = "Jedinica",
	PURCHASE_FOR_FOREIGN_EXCHANGE = "Kupovni za devize",
	MEDIUM_FOR_FOREIGN_EXCHANGE = "Srednji za devize",
	FOREIGN_EXCHANGE_SALES = "Prodajni za devize",
	CURRENCY = "Valuta",
	CURRENCY_CODE = "Šifra valute",
}